refactor AuthStatus tests for resource cleanup
